+++ Changes.md
@@ -1,7 +1,7 @@
-* #771 - updates to the 2019 performance score sheet and add extra space between team info and goals
-
-Release 14.0.0
+Release 14.0.1
 ==============
+* #313 - Use JQuery UI dialog on the score entry page so that the button text can be customized 
+* #771 - updates to the 2019 performance score sheet and add extra space between team info and goals
 * #768 - Confirm zero scores in the subjective web application are not No Shows
 * #769 - Add light grey background to top scores on final scores report to make them easier to see
 * #757 - Add an asterisk next to scores in the final report where a team scored zero on a required goal